Move the updating of the upcoming screen logic to the fragment instead of settings
