+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/UpcomingListMVPView.java
@@ -8,8 +8,6 @@ public interface UpcomingListMVPView { // changed to public because of Kotlin co
 
     void display(List<UpcomingRowViewModel> events);
 
-    void showFirstEvent();
-
     boolean isEmpty();
 
     void askForContactPermission();
